Switch selector scalar to be based off of string, and deal with list in mutation and queries themselves
